refactor: split sample timetables into per-object files under a data package
Each sample timetable (University, Work, School, Conference) now
lives in its own SampleTimetable*.kt file instead of one large
SampleTimetables object, moved under a sample.data subpackage
alongside the shared lesson() helper. SampleTimetables itself is now
a thin dispatcher.

Also adds a new Special Cases sample (renamed from the ad-hoc "Short
Lessons" scenario added earlier on this branch): passing periods,
breaks, and short lessons down to 10 minutes to check the combined/
split time label and priority-based field dropping at small entry
heights, plus a long-titled lesson at two durations to check 2-line
title wrapping.

Verified: ktlintCheck, assembleDebug, and a manual check that the app
installs and every timetable (including the renamed one) still
renders correctly.
