+++ app/src/main/java/de/tobiasschuerg/weekview/sample/ComposeWeekViewActivity.kt
@@ -31,6 +31,7 @@ import de.tobiasschuerg.weekview.compose.WeekViewCompose
 import de.tobiasschuerg.weekview.data.EventConfig
 import de.tobiasschuerg.weekview.data.LocalDateRange
 import de.tobiasschuerg.weekview.data.WeekViewConfig
+import de.tobiasschuerg.weekview.sample.data.SampleTimetables
 import java.time.DayOfWeek
 import java.time.LocalDate
 
@@ -39,12 +40,7 @@ class ComposeWeekViewActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        enableEdgeToEdge(
-            statusBarStyle =
-                SystemBarStyle.dark(
-                    scrim = Color.Transparent.toArgb(),
-                ),
-        )
+        enableEdgeToEdge(statusBarStyle = SystemBarStyle.dark(scrim = Color.Transparent.toArgb()))
 
         val weekViewConfig =
             WeekViewConfig(
